requirements.txt
pyzmq
git+https://github.com/waggle-sensor/pywaggle@v0.37.2
